simple-agents/pyproject.toml
[project]
name = "simple-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "google-generativeai>=0.8.4",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
    "uvicorn>=0.34.0",
]
